% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strand-package.R
\docType{package}
\name{strandai-package}
\alias{strandai}
\alias{strandai-package}
\title{strandai: Client for the Strand Platform API}
\description{
R client for the Strand Platform REST API. Provides functions to upload
Whole Slide Images via GCS resumable upload, submit POSTMAN H&E to
multiplex protein inference jobs, and download OME-Zarr predictions as
SpatialExperiment objects.
}
\section{Quickstart}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{library(strandai)
client <- strand_client()  # reads STRAND_API_KEY
upload <- strand_upload_file(client, "slide.svs")
est <- strand_estimate(client, upload$id, c("CD3", "CD8"))
job <- strand_predict(client, upload$id, c("CD3", "CD8"))
strand_job_wait(job)
spe <- strand_download_results(job)
}\if{html}{\out{</div>}}
}

\section{Authentication}{

Mint API keys at \verb{https://app.strandai.com/settings/api-keys}. Pass with
\code{strand_client(api_key=...)} or set \code{STRAND_API_KEY} in the environment.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://strandai.com}
  \item \url{https://app.strandai.com/docs/api}
}

}
\author{
\strong{Maintainer}: Strand AI \email{engineering@strandai.bio} [copyright holder]

Authors:
\itemize{
  \item Strand AI \email{engineering@strandai.bio} [copyright holder]
}

}
\keyword{internal}
